revert previous leak fixes, fix leak source
leak source is at BroadcastManagerImpl, each register call adds a new
receiver to the Android's LocalBroadcastManager which is static module, while
in unregister we only clear a single receiver, thus leaking the BroadcastManagerImpl
and eventually the activity that its holds in a reference. when solving this
issue the activity will not be leaked, and no need for nulling out references
that may obscure internal objects that are leaked. it's preferable to let
the activity leak in this case, and we catch it by leak canary.