fix leak in CreatePassword/RestoreEnterPassword fragments
TextWatcherAdapter used Timer that did not cancelled and thus caused a leak
(TimerTask uses Thread internally and this thread kept reference to the callback,
to the holding fragment and so on), Timer is not needed at all, and handler can
be used.
As a "release" method is needed on destory to remove handler callbacks,
that requires keeping TextWatcherAdapter refernce, I removed TextWatcherAdapter
and make everything local.