Add new transfer module
The module is responsible for transferring Kin between apps.
At the moment we have implemented only the side that receives a request
to provide the current account public address and returns it back.

The request is received via an intent to a service extending
AccountInfoService. The intent contains information on how to reply back
to the source application that made the request.